move findBoxes to model
