+++ lib/controllers/boxesController.js
@@ -1,9 +1,7 @@
 'use strict';
 
 const
-  { Measurement, Box } = require('../models'),
-  streamTransform = require('stream-transform'),
-  jsonstringify = require('stringify-stream'),
+  { Box } = require('../models'),
   { addCache, clearCache, checkContentType } = require('../helpers/apiUtils'),
   {
     retrieveParameters,
@@ -11,7 +9,6 @@ const
     validateFromToTimeParams,
     checkBoxIdOwner
   } = require('../helpers/userParamHelpers'),
-  { point } = require('@turf/helpers'),
   handleError = require('../helpers/errorHandler');
 
 /**
@@ -126,21 +123,6 @@ const getBoxLocations = function getBoxLocations (req, res, next) {
     });
 };
 
-const geoJsonStringifyReplacer = function geoJsonStringifyReplacer (key, box) {
-  if (key === '') {
-    const coordinates = box.currentLocation.coordinates;
-    box.currentLocation = undefined;
-    box.loc = undefined;
-
-    return point(coordinates, box);
-  }
-
-  return box;
-};
-
-const doNothingStreamTransformFunction = function doNothingStreamTransformFunction (data) {
-  return data;
-};
 
 /**
  * @api {get} /boxes?date=:date&phenomenon=:phenomenon&format=:format Get all senseBoxes
@@ -158,102 +140,15 @@ const doNothingStreamTransformFunction = function doNothingStreamTransformFuncti
  * @apiSampleRequest https://api.opensensemap.org/boxes?date=2015-03-07T02:50Z,2015-04-07T02:50Z&phenomenon=Temperatur
  */
 const getBoxes = function getBoxes (req, res, next) {
-  const { format, phenomenon, fromDate, toDate } = req._userParams;
-
   // content-type is always application/json for this route
   res.header('Content-Type', 'application/json; charset=utf-8');
-  let stringifier;
-  // format
-  if (format === 'json') {
-    stringifier = jsonstringify({ open: '[', close: ']' });
-  } else if (format === 'geojson') {
-    stringifier = jsonstringify({ open: '{"type":"FeatureCollection","features":[', close: ']}' }, geoJsonStringifyReplacer);
-  }
 
-  const query = {};
-
-  // simple string parameters
-  for (const param of ['exposure', 'model', 'grouptag']) {
-    if (req._userParams[param]) {
-      query[param] = { '$in': req._userParams[param] };
-    }
-  }
-
-  let lastMeasurementTransformer = streamTransform(doNothingStreamTransformFunction);
-  const locFieldTransformer = streamTransform(function locFieldTransformer (box) {
-    if (box.currentLocation) {
-      box.loc = [{ geometry: box.currentLocation, type: 'Feature' }];
-    }
-
-    return box;
-  });
-
-  let execStreamBoxesToClientPromise = Promise.resolve(); // no dates specified -> just resolve to exec the boxquery
-
-  if (fromDate && toDate) { // if dates are specified: first query measurments for the dates
-    execStreamBoxesToClientPromise = Measurement.aggregate([
-      {
-        $match: {
-          createdAt: {
-            '$gt': fromDate.toDate(),
-            '$lt': toDate.toDate()
-          }
-        }
-      },
-      {
-        $group: {
-          _id: '$sensor_id',
-          value: { $first: '$value' },
-          createdAt: { $first: '$createdAt' },
-          sensor_id: { $first: '$sensor_id' }
-        },
-      },
-      { $project: { value: 1, createdAt: 1, sensor_id: 1, _id: 0 } }
-    ])
-      .exec()
-      .then(function (measurements) {
-        query['sensors._id'] = {
-          $in: measurements.map(m => m.sensor_id)
-        };
-        if (phenomenon) {
-          query['sensors.title'] = phenomenon;
-        }
-
-        return measurements;
-      });
-  }
-
-  // finally execute the queries
-  execStreamBoxesToClientPromise
-    .then(function (measurements) {
-      if (Array.isArray(measurements)) {
-        let measurementsLength = measurements.length;
-        lastMeasurementTransformer = streamTransform(function (box) {
-          const sensorsLength = box.sensors.length;
-          for (let i = 0; i < measurementsLength; i++) { //iterate measurments
-            for (let j = 0; j < sensorsLength; j++) {
-              if (box.sensors[j]._id.equals(measurements[i].sensor_id)) {
-
-                measurements[i].sensor_id = undefined;
-                box.sensors[j].lastMeasurement = measurements[i];
-                measurements.splice(i, 1);
-                measurementsLength = measurementsLength - 1;
-
-                return box;
-              }
-            }
-          }
-
-          return box;
-        });
-      }
-
-      Box.find(query, Box.BOX_PROPS_FOR_POPULATION)
-        .populate(Box.BOX_SUB_PROPS_FOR_POPULATION)
-        .cursor({ lean: true })
-        .pipe(lastMeasurementTransformer)
-        .pipe(locFieldTransformer) // effects of toJSON must be applied manually for streams
-        .pipe(stringifier)
+  Box.findBoxesLastMeasurements(req._userParams)
+    .then(function (stream) {
+      stream
+        .on('error', function (err) {
+          res.end(`Error: ${err.message}`);
+        })
         .pipe(res);
     })
     .catch(function (err) {
@@ -424,7 +319,6 @@ const postNewBox = function postNewBox (req, res, next) {
  */
 const getSketch = function getSketch (req, res, next) {
   res.header('Content-Type', 'text/plain; charset=utf-8');
-  // Box.findBoxById(boxId, { format })
   Box.findBoxById(req._userParams.boxId, { populate: false, lean: false })
     .then(function (box) {
       res.send(box.getSketch());
