+++ lib/models/box.js
@@ -35,7 +35,9 @@ const { mongoose } = require('../db'),
   ModelError = require('./modelError'),
   Sketcher = require('@sensebox/sketch-templater'),
   fs = require('fs'),
-  { point } = require('@turf/helpers');
+  { point } = require('@turf/helpers'),
+  streamTransform = require('stream-transform'),
+  jsonstringify = require('stringify-stream');
 
 const templateSketcher = new Sketcher(api_measurements_post_domain);
 
@@ -844,12 +846,106 @@ boxSchema.methods.getLocations = function getLocations ({ format, fromDate, toDa
   return locs;
 };
 
+const locFieldTransformerFunction = function locFieldTransformerFunction (box) {
+  if (box.currentLocation) {
+    box.loc = [{ geometry: box.currentLocation, type: 'Feature' }];
+  }
+
+  return box;
+};
+
+const geoJsonStringifyReplacer = function geoJsonStringifyReplacer (key, box) {
+  if (key === '') {
+    const coordinates = box.currentLocation.coordinates;
+    box.currentLocation = undefined;
+    box.loc = undefined;
+
+    return point(coordinates, box);
+  }
+
+  return box;
+};
+
+boxSchema.statics.findBoxesLastMeasurements = function findBoxesLastMeasurements (opts) {
+  const schema = this;
+
+  const { format, phenomenon, fromDate, toDate } = opts,
+    query = {};
+
+  // simple string parameters
+  for (const param of ['exposure', 'model', 'grouptag']) {
+    if (opts[param]) {
+      query[param] = { '$in': opts[param] };
+    }
+  }
+
+  let stringifier;
+  // format
+  if (format === 'json') {
+    stringifier = jsonstringify({ open: '[', close: ']' });
+  } else if (format === 'geojson') {
+    stringifier = jsonstringify({ open: '{"type":"FeatureCollection","features":[', close: ']}' }, geoJsonStringifyReplacer);
+  }
+
+  const locFieldTransformer = streamTransform(locFieldTransformerFunction);
+
+  if (!fromDate && !toDate) {
+    return Promise.resolve(schema.find(query, BOX_PROPS_FOR_POPULATION)
+      .populate(BOX_SUB_PROPS_FOR_POPULATION)
+      .cursor({ lean: true })
+      .pipe(locFieldTransformer) // effects of toJSON must be applied manually for streams
+      .pipe(stringifier)
+    );
+  }
+
+  if (phenomenon) {
+    query['sensors.title'] = phenomenon;
+  }
+
+  return Measurement.findLatestMeasurementsForSensors(fromDate, toDate)
+    .then(function (measurements) {
+      query['sensors._id'] = {
+        $in: measurements.map(m => m.sensor_id)
+      };
+      if (phenomenon) {
+        query['sensors.title'] = phenomenon;
+      }
+
+      let measurementsLength = measurements.length;
+
+      return schema.find(query, BOX_PROPS_FOR_POPULATION)
+        .populate(BOX_SUB_PROPS_FOR_POPULATION)
+        .cursor({ lean: true })
+        .pipe(streamTransform(function (box) {
+          if (box.currentLocation) {
+            box.loc = [{ geometry: box.currentLocation, type: 'Feature' }];
+          }
+
+          for (let i = 0; i < measurementsLength; i++) { //iterate measurments
+            for (const sensor of box.sensors) {
+              if (sensor._id.equals(measurements[i].sensor_id)) {
+
+                measurements[i].sensor_id = undefined;
+                sensor.lastMeasurement = measurements[i];
+                measurements.splice(i, 1);
+                measurementsLength = measurementsLength - 1;
+
+                return box;
+              }
+            }
+          }
+
+          return box;
+        }))
+        .pipe(stringifier);
+    });
+};
+
 // add integrations Schema as box.integrations & register hooks
 integrations.addToSchema(boxSchema);
 
 const boxModel = mongoose.model('Box', boxSchema);
 
-boxModel.BOX_PROPS_FOR_POPULATION = BOX_PROPS_FOR_POPULATION;
 boxModel.BOX_SUB_PROPS_FOR_POPULATION = BOX_SUB_PROPS_FOR_POPULATION;
 boxModel.BOX_VALID_MODELS = sensorLayouts.models;
 boxModel.BOX_VALID_ADDONS = sensorLayouts.addons;
