+++ lib/models/measurement.js
@@ -100,6 +100,29 @@ measurementSchema.statics.hasDecoder = function hasDecoder (contentType) {
   return true;
 };
 
+measurementSchema.statics.findLatestMeasurementsForSensors = function findLatestMeasurementsForSensors (fromDate, toDate) {
+  return this.aggregate([
+    {
+      $match: {
+        createdAt: {
+          '$gt': fromDate.toDate(),
+          '$lt': toDate.toDate()
+        }
+      }
+    },
+    {
+      $group: {
+        _id: '$sensor_id',
+        value: { $first: '$value' },
+        createdAt: { $first: '$createdAt' },
+        sensor_id: { $first: '$sensor_id' }
+      },
+    },
+    { $project: { value: 1, createdAt: 1, sensor_id: 1, _id: 0 } }
+  ])
+    .exec();
+};
+
 const measurementModel = mongoose.model('Measurement', measurementSchema);
 
 module.exports = {
